refactor(ui): extract chat model pickers to sibling file
Move ModelChoiceOptions, ChatModelSwitcher, and NewChatModelPicker (~210
lines) out of the 8k-line components.tsx into chat-model-switcher.tsx,
next to the chat-model-helpers.ts codecs they depend on. Composer imports
them; no behavior change. Prunes the now-unused Select* / groupModelChoices
/ parseModelChoiceValue imports from components.tsx.
